requirements.txt
django>3.2,<3.3
python-decouple==3.6
django-crispy-forms==1.14.0

# Django rest framework
djangorestframework
markdown
django-filter
coreapi
drf-yasg[validation]

# Reformat and linting
black
flake8